# openapi.json
{"openapi":"3.1.0","info":{"title":"Urja Meter API","description":"A clean API service over the Urja Meter Ops portal.","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/meters":{"get":{"summary":"List Meters","operationId":"list_meters_meters_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","description":"Search query for meters","default":"","title":"Q"},"description":"Search query for meters"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeterListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/meters/{meter_id}/energy":{"get":{"summary":"Get Meter Energy","operationId":"get_meter_energy_meters__meter_id__energy_get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnergyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/meters/{meter_id}/consumption":{"get":{"summary":"Get Meter Consumption","operationId":"get_meter_consumption_meters__meter_id__consumption_get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsumptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/meters/{meter_id}/location":{"get":{"summary":"Get Meter Location","operationId":"get_meter_location_meters__meter_id__location_get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ConsumptionReading":{"properties":{"timestamp":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Timestamp"},"consumption_kwh":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Consumption Kwh"}},"type":"object","title":"ConsumptionReading"},"ConsumptionResponse":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"readings":{"items":{"$ref":"#/components/schemas/ConsumptionReading"},"type":"array","title":"Readings"},"total_consumption_kwh":{"type":"number","title":"Total Consumption Kwh","default":0.0}},"type":"object","required":["meter_id"],"title":"ConsumptionResponse"},"EnergyReading":{"properties":{"timestamp":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Timestamp"},"kwh":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Kwh"},"kvarh":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Kvarh"},"volt_r":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Volt R"}},"type":"object","title":"EnergyReading"},"EnergyResponse":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"readings":{"items":{"$ref":"#/components/schemas/EnergyReading"},"type":"array","title":"Readings"}},"type":"object","required":["meter_id"],"title":"EnergyResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"HealthResponse"},"Location":{"properties":{"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"}},"type":"object","required":["latitude","longitude"],"title":"Location"},"LocationResponse":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"location":{"anyOf":[{"$ref":"#/components/schemas/Location"},{"type":"null"}]}},"type":"object","required":["meter_id"],"title":"LocationResponse"},"Meter":{"properties":{"meter_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Meter Id"},"serial_no":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Serial No"},"make":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Make"},"phase_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phase Type"},"install_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Install Status"},"dt_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dt Code"}},"type":"object","title":"Meter"},"MeterListResponse":{"properties":{"meters":{"items":{"$ref":"#/components/schemas/Meter"},"type":"array","title":"Meters"},"total":{"type":"integer","title":"Total","default":0},"page":{"type":"integer","title":"Page","default":1},"page_size":{"type":"integer","title":"Page Size","default":0}},"type":"object","title":"MeterListResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}